+++ i18n/messages.no.js
@@ -13,6 +13,7 @@ window.ParsleyConfig = window.ParsleyConfig || {};
           , digits:     "Denne verdien må være tall."
           , dateIso:    "Denne verdien må være en gyldig dato (YYYY-MM-DD)."
           , alphanum:   "Denne verdien må være alfanumerisk(tall eller bokstaver)."
+          , phone:      "Denne verdien må være et gyldig telefonnummer."
         }
       , notnull:        "Denne verdien kan ikke være tom."
       , notblank:       "Denne verdien kan ikke være blank."
@@ -24,9 +25,20 @@ window.ParsleyConfig = window.ParsleyConfig || {};
       , minlength:      "Denne verdien er for kort. Den må være minst %s tegn."
       , maxlength:      "Denne verdien er for lang. Den må ikke være lenger enn %s tegn."
       , rangelength:    "Denne verdien har feil lengde. Lengden må være mellom %s og %s tegn."
+      , mincheck:       "Du må velge minst %s alternativer."
+      , maxcheck:       "Du kan ikke velge mer enn %s alternativer."
+      , rangecheck:     "Du må velge mellom %s og %s alternativer."
       , equalto:        "Denne verdien må være lik."
 
       // parsley.extend ///////////////////////////////
+       , minwords:       "Denne verdien må inneholde minst %s ord."
+       , maxwords:       "Denne verdien kan ikke inneholde mer enn %s ord."
+       , rangewords:     "Denne verdien må ha mellom %s og %s ord."
+       , greaterthan:    "Denne verdien må være større enn %s."
+       , lessthan:       "Denne verdien må være mindre enn %s."
+       , beforedate:     "Datoen må være før %s."
+       , afterdate:      "Datoen må være etter %s."
+       , americandate:   "Datoen må være på gyldig format (MM/DD/YYYY)."
     }
   });
 }(window.jQuery || window.Zepto));